Refactor: move LeaderState to leader_state.rs; add runtime.rs, move runtime impl to this file
