+++ openraft/src/core/mod.rs
@@ -11,9 +11,9 @@ mod raft_core;
 pub(crate) mod replication;
 #[cfg(test)]
 mod replication_state_test;
+mod runtime;
 mod vote;
 
-use std::collections::BTreeMap;
 use std::collections::BTreeSet;
 use std::fmt::Debug;
 use std::sync::Arc;
@@ -21,6 +21,7 @@ use std::sync::Arc;
 pub use effective_membership::EffectiveMembership;
 use futures::future::AbortHandle;
 use futures::future::Abortable;
+use leader_state::LeaderState;
 use maplit::btreeset;
 use rand::thread_rng;
 use rand::Rng;
@@ -42,7 +43,6 @@ use crate::config::Config;
 use crate::config::SnapshotPolicy;
 use crate::engine::Engine;
 use crate::error::AddLearnerError;
-use crate::error::ClientWriteError;
 use crate::error::ExtractFatal;
 use crate::error::Fatal;
 use crate::error::ForwardToLeader;
@@ -51,18 +51,15 @@ use crate::error::NotAllowed;
 use crate::metrics::RaftMetrics;
 use crate::metrics::ReplicationMetrics;
 use crate::raft::AddLearnerResponse;
-use crate::raft::ClientWriteResponse;
 use crate::raft::RaftMsg;
 use crate::raft::RaftRespTx;
 use crate::raft::VoteResponse;
 use crate::raft_types::LogIdOptionExt;
-use crate::replication::ReplicaEvent;
 use crate::replication::ReplicationStream;
 use crate::storage::RaftSnapshotBuilder;
 use crate::versioned::Versioned;
 use crate::vote::Vote;
 use crate::Entry;
-use crate::EntryPayload;
 use crate::LogId;
 use crate::MessageSummary;
 use crate::Node;
@@ -658,171 +655,8 @@ impl ServerState {
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 
-/// Volatile state specific to the Raft leader.
-struct LeaderState<'a, C: RaftTypeConfig, N: RaftNetworkFactory<C>, S: RaftStorage<C>> {
-    pub(super) core: &'a mut RaftCore<C, N, S>,
-
-    /// A mapping of node IDs the replication state of the target node.
-    pub(super) nodes: BTreeMap<C::NodeId, ReplicationState<C>>,
-
-    /// The metrics about a leader
-    pub replication_metrics: Versioned<ReplicationMetrics<C::NodeId>>,
-
-    /// The stream of events coming from replication streams.
-    pub(super) replication_rx: mpsc::UnboundedReceiver<(ReplicaEvent<C, S::SnapshotData>, Span)>,
-
-    /// The cloneable sender channel for replication stream events.
-    pub(super) replication_tx: mpsc::UnboundedSender<(ReplicaEvent<C, S::SnapshotData>, Span)>,
-
-    /// Channels to send result back to client when logs are committed.
-    pub(super) client_resp_channels: BTreeMap<u64, RaftRespTx<ClientWriteResponse<C>, ClientWriteError<C::NodeId>>>,
-}
-
-impl<'a, C: RaftTypeConfig, N: RaftNetworkFactory<C>, S: RaftStorage<C>> MetricsProvider<C::NodeId>
-    for LeaderState<'a, C, N, S>
-{
-    fn get_leader_metrics(&self) -> Option<&Versioned<ReplicationMetrics<C::NodeId>>> {
-        Some(&self.replication_metrics)
-    }
-}
-
-impl<'a, C: RaftTypeConfig, N: RaftNetworkFactory<C>, S: RaftStorage<C>> LeaderState<'a, C, N, S> {
-    /// Create a new instance.
-    pub(self) fn new(core: &'a mut RaftCore<C, N, S>) -> Self {
-        let (replication_tx, replication_rx) = mpsc::unbounded_channel();
-        Self {
-            core,
-            nodes: BTreeMap::new(),
-            replication_metrics: Versioned::new(ReplicationMetrics::default()),
-            replication_tx,
-            replication_rx,
-            client_resp_channels: Default::default(),
-        }
-    }
-
-    /// Transition to the Raft leader state.
-    #[tracing::instrument(level="debug", skip(self), fields(id=display(self.core.id), raft_state="leader"))]
-    pub(self) async fn run(mut self) -> Result<(), Fatal<C::NodeId>> {
-        // Setup state as leader.
-        self.core.last_heartbeat = None;
-        self.core.next_election_timeout = None;
-        self.core.engine.state.vote.commit();
-
-        // Spawn replication streams for followers and learners.
-        let targets = self
-            .core
-            .engine
-            .state
-            .effective_membership
-            .node_ids()
-            .filter(|elem| *elem != &self.core.id)
-            .cloned()
-            .collect::<Vec<_>>();
-
-        for target in targets {
-            let state = self.spawn_replication_stream(target, None).await;
-            self.nodes.insert(target, state);
-        }
-
-        // Commit the initial entry when new leader established.
-        self.write_entry(EntryPayload::Blank, None).await?;
-
-        self.leader_loop().await?;
-
-        Ok(())
-    }
-
-    #[tracing::instrument(level="debug", skip(self), fields(id=display(self.core.id)))]
-    pub(self) async fn leader_loop(mut self) -> Result<(), Fatal<C::NodeId>> {
-        // report the leader metrics every time there came to a new leader
-        // if not `report_metrics` before the leader loop, the leader metrics may not be updated cause no coming event.
-        self.core.report_metrics(Update::Update(Some(self.replication_metrics.clone())));
-
-        loop {
-            if !self.core.engine.state.server_state.is_leader() {
-                tracing::info!(
-                    "id={} state becomes: {:?}",
-                    self.core.id,
-                    self.core.engine.state.server_state
-                );
-
-                // implicit drop replication_rx
-                // notify to all nodes DO NOT send replication event any more.
-                return Ok(());
-            }
-
-            self.core.report_metrics_if_needed(&self);
-            self.core.engine.metrics_flags.reset();
-
-            tokio::select! {
-                Some((msg,span)) = self.core.rx_api.recv() => {
-                    self.handle_msg(msg).instrument(span).await?;
-                },
-
-                Some(update) = self.core.rx_compaction.recv() => {
-                    tracing::info!("leader recv from rx_compaction: {:?}", update);
-                    self.core.update_snapshot_state(update);
-                }
-
-                Some((event, span)) = self.replication_rx.recv() => {
-                    tracing::info!("leader recv from replication_rx: {:?}", event.summary());
-                    self.handle_replica_event(event).instrument(span).await?;
-                }
-
-                Ok(_) = &mut self.core.rx_shutdown => {
-                    tracing::info!("leader recv from rx_shudown");
-                    self.core.set_target_state(ServerState::Shutdown);
-                }
-            }
-        }
-    }
-
-    #[tracing::instrument(level = "debug", skip(self, msg), fields(state = "leader", id=display(self.core.id)))]
-    pub async fn handle_msg(&mut self, msg: RaftMsg<C, N, S>) -> Result<(), Fatal<C::NodeId>> {
-        tracing::debug!("recv from rx_api: {}", msg.summary());
-
-        match msg {
-            RaftMsg::AppendEntries { rpc, tx } => {
-                let res = self.core.handle_append_entries_request(rpc).await.extract_fatal()?;
-                let _ = tx.send(res);
-            }
-            RaftMsg::RequestVote { rpc, tx } => {
-                let res = self.core.handle_vote_request(rpc).await.extract_fatal()?;
-                let _ = tx.send(res);
-            }
-            RaftMsg::InstallSnapshot { rpc, tx } => {
-                let res = self.core.handle_install_snapshot_request(rpc).await.extract_fatal()?;
-                let _ = tx.send(res);
-            }
-            RaftMsg::CheckIsLeaderRequest { tx } => {
-                self.handle_check_is_leader_request(tx).await;
-            }
-            RaftMsg::ClientWriteRequest { rpc, tx } => {
-                self.write_entry(rpc.payload, Some(tx)).await?;
-            }
-            RaftMsg::Initialize { tx, .. } => {
-                self.core.reject_init_with_config(tx);
-            }
-            RaftMsg::AddLearner { id, node, tx, blocking } => {
-                self.add_learner(id, node, tx, blocking).await;
-            }
-            RaftMsg::ChangeMembership {
-                members,
-                blocking,
-                turn_to_learner,
-                tx,
-            } => self.change_membership(members, blocking, turn_to_learner, tx).await?,
-            RaftMsg::ExternalRequest { req } => {
-                req(ServerState::Leader, &mut self.core.storage, &mut self.core.network);
-            }
-        };
-
-        Ok(())
-    }
-}
-
 /// A struct tracking the state of a replication stream from the perspective of the Raft actor.
-struct ReplicationState<C: RaftTypeConfig> {
+pub(crate) struct ReplicationState<C: RaftTypeConfig> {
     pub matched: Option<LogId<C::NodeId>>,
     pub remove_since: Option<u64>,
     pub repl_stream: ReplicationStream<C>,
